Upgrade to Vue 3
This commit upgrades all Vue components to be Vue 3-based, using the
Composition API. As part of this, Vuex is replaced by Pinia, Webpack by
Vite, Node/NPM bumped, and some changes have been made to existing
components.

The auto-save indicator now uses a single prop and events are passed
through Vue. `django-channels` is also replaced with a native solution
as the package is quite outdated.